updated tests to make sure the right guest and admin permissions are created, and some tidying up and reorganising
