+++ db/seeds/example_data/example_data.seeds.rb
@@ -1,3 +1,3 @@
 # Use the new centralized ExampleDataSeeder class
-seeder = Seek::ExampleDataSeeder.new
+seeder = Seek::ExampleData::ExampleDataSeeder.new
 seeder.seed_all
